requirements.txt
python-barcode==0.9.0
zklib==0.1.1
